Move the migrate and webhook code into the other directory
